Improve elrrsync test package names.  Delete JUnit tests that mock what's being tested.
