--- src/test/java/com/deloitte/elrr/datasync/service/package-info.java
@@ -1,5 +1,0 @@
-/**
- * @author mnelakurti
- *
- */
-package com.deloitte.elrr.datasync.service;
