--- src/test/java/com/deloitte/elrr/datasync/util/MockLRSDataTest.java
@@ -1,27 +1,0 @@
-/**
- *
- */
-package com.deloitte.elrr.datasync.util;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-/**
- * @author mnelakurti
- *
- */
-@ExtendWith(MockitoExtension.class)
-class MockLRSDataTest {
-
-    /**
-     *
-     */
-    @Test
-    void test() {
-        assertNotNull(MockLRSData.getLearnerStatements());
-    }
-
-}
