+++ src/test/java/com/deloitte/elrr/test/datasync/test/entity/AuditableTest.java
@@ -1,0 +1,21 @@
+/** */
+package com.deloitte.elrr.test.datasync.test.entity;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import com.deloitte.elrr.datasync.entity.Auditable;
+import com.deloitte.elrr.test.datasync.test.util.ValueObjectTestUtility;
+
+/**
+ * @author mnelakurti
+ */
+@ExtendWith(MockitoExtension.class)
+class AuditableTest {
+
+  @Test
+  void test() {
+    ValueObjectTestUtility.validateAccessors(Auditable.class);
+  }
+}
