+++ src/test/java/com/deloitte/elrr/test/datasync/test/util/ValueObjectTestUtility.java
@@ -1,4 +1,4 @@
-package com.deloitte.elrr.datasync.util;
+package com.deloitte.elrr.test.datasync.test.util;
 
 import com.openpojo.reflection.impl.PojoClassFactory;
 import com.openpojo.validation.Validator;
